requirements.txt
streamlit
deep-translator
python-docx
pdf2docx